[Landing Page] Added opaque field photograph
Received request to include the memorial field on the homepage as an opaque backdrop. Executed successfully on desktop and hidden on mobile. On mobile BP it conflicted with the footer (regardless of z-index) and the mobile navigation; image is not recognizable on mobile so hiding it was the easiest solution.